:truck: (interface): Rename file interface/CoreJPEG.hpp to interface/JPEG.hpp
