+++ drivers/CoreVideo/include/CoreVideo.hpp
@@ -7,7 +7,7 @@
 
 #include "LKCoreSTM32HalBase.h"
 #include "interface/CoreGraphics.hpp"
-#include "interface/CoreJPEG.hpp"
+#include "interface/JPEG.hpp"
 #include "interface/CoreSDRAM.hpp"
 #include "interface/DMA2D.hpp"
 #include "interface/DSI.hpp"
